requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
crispy-bootstrap5==0.6
Django==3.2.9
django-bootstrap-v5==1.0.7
django-crispy-forms==1.13.0
django-extensions==3.1.5
django-shapeshifter==18.9.23
django-widget-tweaks==1.4.9
importlib-metadata==3.10.1
Pillow==8.4.0
pytz==2021.3
reportlab==3.6.2
soupsieve==2.3.1
sqlparse==0.4.2
zipp==3.6.0
